Move footer and it's test to a dedicated Footer folder
