+++ client/src/components/Footer/Footer.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
-import Footer from "./src/components/Footer";
+import Footer from "./Footer";
 
 describe("Footer Component", () => {
   it("renders a link", () => {
